+++ Mochi.DearImGui.OpenTK/PlatformBackend.cs
@@ -73,7 +73,7 @@ public unsafe sealed partial class PlatformBackend : IDisposable
         {
             string name = GetType().FullName ?? nameof(PlatformBackend);
             int nameByteCount = Encoding.UTF8.GetByteCount(name) + 1;
-            byte* nameP = (byte*)ImGui.MemAlloc((ulong)nameByteCount);
+            byte* nameP = (byte*)ImGui.MemAlloc((nuint)nameByteCount);
             Io->BackendPlatformName = nameP;
             Span<byte> nameSpan = new(nameP, nameByteCount);
             int encodedByteCount = Encoding.UTF8.GetBytes(name.AsSpan().Slice(0, nameSpan.Length - 1), nameSpan);
